Avoids base64 decoding token in scheme handler
Some tokens, like JWTs are technically b64, but actually split up into 3
parts of b64, separated by ".". In other words, the entire token isn't
decoded into base64, the parts are. This is done by the token parser.
The scheme handler will just pass along the token to the backend for the
decoding.